Feat/notification : update file migration table notification  and update name function in repo notification

+++ internal/service/notification.go
@@ -1,6 +1,8 @@
 package service
 
 import (
+	"errors"
+
 	"github.com/Giafn/Depublic/internal/entity"
 	"github.com/Giafn/Depublic/internal/repository"
 	"github.com/google/uuid"
@@ -10,7 +12,7 @@ type NotificationService interface {
 	FindAllNotification(userID uuid.UUID) ([]entity.Notification, error)
 	FindNotificationByID(notificationID uuid.UUID) (*entity.Notification, error)
 	CreateNotification(notification *entity.Notification) (*entity.Notification, error)
-	UpdateSeenAllNotification(userID uuid.UUID) ([]entity.Notification, error)
+	MarkAllNotificationsAsSeen(userID uuid.UUID) ([]entity.Notification, error)
 	DeleteNotification(notificationID uuid.UUID) (bool, error)
 	DeleteSeenNotification(userID uuid.UUID) (bool, error)
 }
@@ -37,14 +39,25 @@ func (s *notificationService) CreateNotification(notification *entity.Notificati
 	return s.NotificationRepo.CreateNotification(notification)
 }
 
-func (s *notificationService) UpdateSeenAllNotification(userID uuid.UUID) ([]entity.Notification, error) {
-	return s.NotificationRepo.UpdateSeenAllNotification(userID)
+func (s *notificationService) MarkAllNotificationsAsSeen(userID uuid.UUID) ([]entity.Notification, error) {
+	data, err := s.NotificationRepo.FindNotificationSeen(userID)
+
+	if err != nil {
+		return nil, err
+	}
+
+	if len(data) == 0 {
+		return nil, errors.New("Tidak ada notifikasi yang sudah dilihat")
+	}
+
+
+	return s.NotificationRepo.MarkAllNotificationsAsSeen(userID)
 }
 
 func (s *notificationService) DeleteNotification(notificationID uuid.UUID) (bool, error) {
     notification, err := s.NotificationRepo.FindNotificationByID(notificationID)
 
-	if err != nil {
+	if err != nil || notification == nil {
 		return false, err
 	}
 
@@ -52,6 +65,16 @@ func (s *notificationService) DeleteNotification(notificationID uuid.UUID) (bool
 }
 
 func (s *notificationService) DeleteSeenNotification(userID uuid.UUID) (bool, error) {
+	data, err := s.NotificationRepo.FindNotificationSeen(userID)
+
+	if err != nil {
+		return false, err
+	}
+
+	if len(data) == 0 {
+		return false, errors.New("Tidak ada notifikasi yang sudah dilihat")
+	}
+
 	return s.NotificationRepo.DeleteSeenAllNotification(userID)
 }
